(PC-2845): Finish migration and delete previous one for redundancy
